[MINOR] solve comments
